+++ dsh_sdk/src/error.rs
@@ -37,7 +37,6 @@ pub enum DshError {
     #[cfg(feature = "bootstrap")]
     #[error("Error getting group id, index out of bounds for {0}")]
     IndexGroupIdError(crate::dsh::datastream::GroupType),
-    #[cfg(feature = "bootstrap")]
     #[error("No tenant name found")]
     NoTenantName,
     #[cfg(feature = "bootstrap")]
@@ -52,4 +51,21 @@ pub enum DshError {
     #[cfg(feature = "metrics")]
     #[error("Hyper error: {0}")]
     HyperError(#[from] hyper::http::Error),
+}
+
+#[cfg(feature = "rest-token-fetcher")]
+#[derive(Error, Debug)]
+#[non_exhaustive]
+pub enum DshRestTokenError {
+    #[error("Client ID is unknown")]
+    UnknownClientId,
+    #[error("Client secret not set")]
+    UnknownClientSecret,
+    #[error("Unexpected failure while fetching token from server: {0}")]
+    FailureTokenFetch(reqwest::Error),
+    #[error("Unexpected status code: {status_code}, error body: {error_body:#?}")]
+    StatusCode {
+        status_code: reqwest::StatusCode,
+        error_body: reqwest::Response,
+    },
 }