+++ dsh_sdk/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Add token fetcher for DSH REST API
+- Add platform enum with metadata
+- Add example on how to use the token fetcher and rest api client
+
+
 ## [0.4.1] - 2024-06-13
 ### Added
 - Add loading Certificates and Keys from $PKI_CONFIG_DIR